Rename read_trace.h to reader.h
